feat(units): Fahd UoM rollout — pack units, decimals, purchasing UX, outage fix
Units of measure, end to end:
- ProductUnitBarcode: packaging (carton) EANs on units; resolved by the
  POS/purchasing scan flows, product search, and the price checker;
  preserved across product edits (absent-key keeps stored codes)
- fahd migration: CAR_PART_D2 pack codes become ProductUnits via a
  sale-evidence classifier (piece-alias codes keep ringing pieces);
  upgrade path repairs already-migrated servers in place
  (deploy/onprem/backfill-fahd-units.sh = backfill + 29 confirmed
  per-product fixups incl. fractional egg trays and the x360 carton);
  set_product_unit management command for one-off unit fixups
- decimal quantities in purchasing end to end (order/receive/adjust,
  3dp wire format) and fractional entry in POS, gated per unit
- purchasing UX parity with POS: unit-aware last cost that rescales on
  unit switch, tap-to-select draft lines with typed-quantity preview,
  decimal quick-typing after scan/tap, stable line order (merge in
  place), one-off order discount under the landed costs
  (extra_discount_amount, folded into totals + preview)
- product EDIT flow gets the same units section as create; PO Edit now
  available for submitted-but-untouched orders (expected stock rebuilt)

Ops:
- interval outages root-caused to uvicorn --limit-max-requests worker
  recycling (all workers died together under steady polling); recycling
  now defaults off; fix-backend-outages.{ps1,sh} repair live installs,
  disable-watchdog.{ps1,sh} for opting out of the autostart tasks
- frontend analytics ingestion batches aggressively: 120s cadence,
  full-queue drain in 100-event chunks, rate-limited immediate flushes

+++ frontend/lib/dev/purchasing_preview.dart
@@ -273,9 +273,9 @@ PurchaseOrderLine _line({
   required int id,
   required String name,
   String? sku,
-  required int quantity,
-  int received = 0,
-  int damaged = 0,
+  required double quantity,
+  double received = 0,
+  double damaged = 0,
   required double unitCost,
 }) {
   final open = quantity - received - damaged;
